VideoKyc, Camera, Bussiness, SupersetCompo and css changes
